[mojo-lang] Introduce `lit.owned.arg.decl` op to model owned memory args. (#11681)
This introduces a new op (and makes the parser generate it) that
captures the destructor associated with an owned in-memory argument
to a function.  This also provides a uniform representation for
the lifetime pass to work on.

MODULAR_ORIG_COMMIT_REV_ID: b0d0d04c246be65d8fb26cf2044607bc80146091