[mojo-lang] Build out the first steps of the lifetime checker. (#11718)
This builds the core data structures for modeling lifetime tracked
values, collects them and walks the IR in a reasonable way, and
implements a super-early sketch of pass two, which diagnoses uses
of uninitialized values (currently done by Mem2Reg, but that will
eventually be removed because this check needs to be done early).

This code is still incomplete in a bunch of ways, notably not
handling control flow and all the other ops, and not being field
sensitive.  I'll phase these features in over time, and this code
is thus hard-gated to only applying to types with `__del___`
destructors for now.

MODULAR_ORIG_COMMIT_REV_ID: b3a2368ee79d7a7d4979f58b15f173a2af3c1080