[Kernels] Fix bug in packed B matmul and re-enable tests (#11729)
There was an incorrect assumption in the get_tile function in the
BTileGenerator that tile_k would be constant for every call. However,
this is not the case for the residual K tile, since the caller adjusts
tile_k to be the minimum of the K bound and the original tile_k. This
caused an incorrect offset calculation to the tile in the packed buffer.

This only showed up in debug builds because the B tile size is 1KB
instead of 128KB in debug builds. This means that in release builds K
always fit within a single tile for the existing test cases. Added a
test case where K does not fit within in single tile for release builds,
so this would have been caught in CI.

LINALG_ORIG_REV_ID: 6e519cf7b79c1f5683347d3d305cb4f5965de68e