[KGEN] Include the BuildInfoAttr in the mlir module (#11734)
This includes the BuildInfoAttr as an attr in the module. This makes
KGEN honor the build configuration of the kernels (as well as the
profile level). For now the build info attr cannot be specified by the
user and is instead picked up from the current build configuration.

MODULAR_ORIG_COMMIT_REV_ID: d21f949f5ff071735cba254e6b8230f755ade2d3