[GenericML] Load drivers from Cache (#11612)
This is a second attempt after reverting -
[Internal link]

The previous PR was reverted because it was creating too many temporary
files. Majority of the logic is the same. Main difference is instead of
unique files we will use a Lock to avoid duplication. This slows down
test run a bit, but doesn't seem to be a significant slowdown.

MODULAR_ORIG_COMMIT_REV_ID: 68483f12152c70b3c8e9a316a8ad74884b17c1ad