[Kernels] Fix bug in packed matmul with transposed B (#11839)
A mistake in the loop ordering in pack_b meant that the tiles were being
stored in col major order when B was transposed. The packed matmul impl
expects the tiles to be in row major order regardless of if B was
transposed before packing.

This only showed up in debug builds because the B tile size is 1KB
instead of 128KB in release builds. This meant that there were no
existing test cases where both n > tile_n and k > tile_k, which is where
this bug shows up. Therefore, added a test case where this should be
true, although it is a bit brittle because it depends on some of the
matmul internals.

LINALG_ORIG_REV_ID: 6c7c81230fef4a03a84ea709acf74ce4b22d0c6c