[mojo] Add a flag for doc validation to mojo (#11764)
This allows for users to control doc string validation from mojo,
which is intended to be the main driver for mojo code. Makes sense
for it to be able to toggle this behavior.

This also renames the option to `doc-validate`, which makes it a bit
easier to have additional doc related options (like `doc-warn-blah`).

MODULAR_ORIG_COMMIT_REV_ID: ba83945f018c2334287147af75c45dcc4d4eb6ac