[mojo-lang][NFC] Add a new MojoParserConfig to simplify importMojoFile (#11765)
We're starting to collect lots of options that configure parsing a Mojo
file. This PR moves the various options into a new config to make it
a bit nicer to parse a Mojo file, and to make it easier to add new config
options in the future.

MODULAR_ORIG_COMMIT_REV_ID: e4672358b193b65cbcdd813c028115e15e97c720