[mojo-notebook] Implement a basic Jupyter Kernel for Mojo (#11739)
The kernel is split into two different parts. The first is a python kernel
that handles communication/registration/etc. with the jupyter client.
The second is a C++ kernel library that handles communicating with
LLDB and running the repl. The C++ kernel library exports a simplified
C api to enable communication between the two.

This kernel provides enough for us to start executing notebooks, but
still lacks various niceties that we'll want with the jupyter experience.
Some of this is on the kernel side, but a lot is on the LLDB side. As
we add more features to the LLDB expression handler, they should
get automatically picked up by the notebook kernel (given that the kernel
is mostly agnostic of the Mojo plugin, and interacts with the generic
LLDB API).

MODULAR_ORIG_COMMIT_REV_ID: 4dc2692e71a0cf3c0d0d6d1080c741eb9148b733