[KGEN] Don't loop infinitely in `getAutoImportPath`. (#11898)
If we aren't in the modular tree when we run `mojo`/`kgen` we'll hit an
infinite loop. This is because we search up the current path to find the
modular directory, and if we aren't rooted in a modular directory, we'll
hit `path = path.parent_path()` infinitely. This is problematic because
`"/".parent_path() == "/"`.

MODULAR_ORIG_COMMIT_REV_ID: b20b1442498953e9b3dee20c365ceeb3b3f5aa8b