[KGEN] Fix bozo bug in CheckLifetimes. (#11937)
CheckLifetimes had a trivial bug I introduced that caused it to never
generated diagnostics. This was unnoticed because the
-verify-diagnostics
flag was checking for expected diagnostics from the wrong file (a
generated
tmp file).

Fix the bug, and fix the check line.

While here, change the diagnostic emission to emit at the location of an
operation instead of the operation itself. We don't want MLIR to
helpfully
tell us what the MLIR representation of some IR is in a user-facing
diagnostic!

MODULAR_ORIG_COMMIT_REV_ID: 710edb6bd87e8d9d562624a93f2ca5042025ddb2