[KGEN][CheckLifetimes] Add decl names to diagnostics. (#11943)
This improves the specificity of the uninit diagnostics by mentioning
the name of the decl in question.  This runs into a minor missing thing
in KGEN, tracked in Issue #11918 for closure value names.

As part of doing this, we refactor stuff out to a new LifetimeTrackable
helper struct, which reduces the amount of stuff we need to track in
ValueInfo.

MODULAR_ORIG_COMMIT_REV_ID: a9fdc15aa49f42d88014c74ae6399108ae5a7dce