[mojo-lang] Rearrange struct dtor processing logic a bit. (#11968)
This changes it to look up the destructor at the end of struct body
type checking, and to use an existing helper.  This also changes
StructDeclOp to maintain a typed attribute instead of just a symbol,
as this allows us to typecheck when emitting a call to the dtor.

MODULAR_ORIG_COMMIT_REV_ID: 792e97d53077c13ee39c9b139e48aa23b171ffb8