[MBlack] Result parameters in call expressions (#11994)
The following is valid Mojo, but currently fails to parse with MBlack:

```
f[Int, VariadicList[Int](1, 2, 3)->out]()
```

That's because the grammar treats `[...]` as a subscript, and does not
account for `->` appearing in that context. Update the grammar to
support this. Now the above code is formatted as:

```
f[Int, VariadicList[Int](1, 2, 3) -> out]()
```

MODULAR_ORIG_COMMIT_REV_ID: c3c12a51780f3fed9990794af3cd4d341a413e07