[Cache][GenericML] Add option to give preference when looking for hardware specific drivers. (#11930)
There are higher chances of cache misses if we only consider Host
Specific Keys when looking for drivers. This PR adds ability to provide
keys to backups in cases of cache misses.

We will look for a cache artifact that's host specific, and if it fails
we will relax threading requirements in host info and look for an object
that satisfies this requirement.

Also add the ability to cache-mgr to add objects to cache with keys
wrapped in StaticHostInfo.

MODULAR_ORIG_COMMIT_REV_ID: 4227b952288eef9c467318d42b069fd098499ce0