[mojo-lang] When the source file is unknown, use the MainFileID. (#11998)
In the REPL, we see that the buffer location on SMDiagnostics are incorrect, even though the SMFixIt ranges are correct. This led back to the fact that mojo parsing unconditionally turns SMLoc into mlir::Location, and uses an "<unknown>" file name when there's no name on the buffer in the SourceMgr. We can detect this when we're translating *back* to an SMLoc, and correctly emit the SMLoc for use by the REPL (and others).

MODULAR_ORIG_COMMIT_REV_ID: c78a4b1756f0657ea457414858aeb7a27dfb947c