[KGEN][CheckLifetimes] First crack at destructor insertion. (#12059)
Now that we know that all values are defined before use, we can
use a backward scan over the IR to insert destructors.  Whenever
we see the last use of a value, we emit a destructor call for the
value after the use.  This ensures we get predictable early
destruction of a value.

This patch has a number of limitations, but is a promising step

MODULAR_ORIG_COMMIT_REV_ID: 2642d4b8350394d66c3ca328c3bb19292e0d5c1f