[Kernels] Add epilog closure to matmul and use for FC kernel (#12058)
This PR adds an optional epilog closure to the legacy (non-generated)
matmul, which is used for a fully connected kernel, and will also be
used for activations in the future. Note that the epilog needs to be a
runtime closure, not a compile time closure like it is in the generated
matmul. This is because making the epilog a compile-time closure would
force all of the methods in the TiledMatmul to be @always_inline, which
degrades performance by 5% (not sure why - maybe ICache pollution).

The PR also makes the legacy matmul the default for the fully connected
kernel. This is because the generated matmul does not supported
pre-packing B yet and it is still 4-5% less performant than the legacy
matmul.

LINALG_ORIG_REV_ID: 21379fa57605e739a07fe3d22e7b7c9e860ccc80