[Mojo] OutputChain to mediate between MEF and Mojo async calls (#12128)
This promotes the OutputChain alias to a full class, containing:
 - the true AsyncValueRef<Chain>
 - a location for errors
 - refs to AsyncValues to keep alive
 - an optional profiling entry

The Mojo OutputChainPtr and CompilerRT interface are unchanged.

The only semantic change is that the AsyncValue refs and profiling entry
will be closed out *before* the AsyncValueRef<Chain> is emplaced. In
this way profiling ends as soon as the Mojo kernel declares it is done,
and downstream waiters will not see that any of the refs are still
alive.

With this in place it will be straightforward to associate sensible
profiling entry labels to sub-tasks, which I'll do in a follow up.

Perhaps OutputChain is better called MojoAsyncCallContext or something
more specific.

MODULAR_ORIG_COMMIT_REV_ID: c238ad87cdfd81191a3af99845205bcb85ea40b3