[Kernels] Add initial scaffolding for matmul/FC + activation, NFC (#12147)
Adds the initial scaffolding to add an activation epilogue to the tiled
matmul.

Will add tests and actually wire in the activations from the kernel
shims in a followup PR.

This does not support parametric ReLU (PRELU), since that activation
needs to capture a runtime value. That might require a separate shim for
matmul+prelu and fc+prelu to pass in the PRELU alpha argument. Deferring
this since passing that runtime value will probably become easier once
the matmul is jit'ed and the activation can be passed via a closure to
the kernel.

NN_ORIG_REV_ID: dc708811a7a98b6a2809e09d58cfc0d239b01098