[mojo-lldb] Apply multiple fix-its. (#12145)
This patch takes advantage of the fact that fixits are ordered by start
range to apply multiple of them to an expression. This has the potential
to be strange/problematic for really complex expressions, but in those
cases we're likely in trouble anyway.

The following example:
```
 3
 fn foo():
   4
   print(32)
 foo
```
is fixed up to:
```
_ = 3
fn foo():
  _ = 4
  print(32)
foo()
```

and prints 32 as expected.

MODULAR_ORIG_COMMIT_REV_ID: 8c7458caa7985f4a2534819d1234dadd844a37f8