[mojo-lldb] Add initial support for persistent variables (#12022)
The general structure in this PR is that the entry expression function is
now passed a reference to a struct containing pointers holding the addresses
of previous variables. New persistent variables are added (in REPL mode) by
processing all of the variables at the top-level of the expression function, rewriting
them to refer to fields within the input struct instead of local values. This is semi tricky
to do in the current state of Mojo given that we don't have global variables. For now
we malloc space for the variable and use that in place of a local allocation. We'll want
to evolve this as we both build out more type system support in LLDB, and add
features like global variables in lit.

This PR also only handles adding new variables, it doesn't yet handle removing
previous ones (e.g. as needed when re-executing a previous cell in jupyter).

MODULAR_ORIG_COMMIT_REV_ID: cc5ffdd2f2efe3bbbefc7c9f5f5261dfdc032779