[mojo-lldb] Pull in IRExecutionUnit and LLVMUserExpression as hack to own (#12155)
We currently use these to perform JIT codegen for Mojo expressions, but
we're running into various issues where it makes sense to pull these in
and mutate.

A major issue that we're facing right now is that we dynamically
link to liblldb (which has its own copy of LLVM) and have one statically linked
into the plugin. This is currently causing crashes during codegen where LLVM
IR on the Mojo side references globals that differ from those expected on the
LLDB side. By having all of the codegen processed owned by the plugin, we
can safely sandbox references to LLVM state, and purely use JIT addresses
on the LLDB boundary.

A major seconday benefit is that this also lets us have greater control over
the JIT process. LLVMUserExpression currently uses MCJIT to perform compilation
which aside from being ancient, completely differs from all of the JIT infrastructure
that we've built on the Mojo side, meaning we don't benefit from a lot of what
we've built and end up reimplementing it (with less features). In addition to infra
reuse, this will also make it much easier to iterate on features necessary for a good
repl experience (e.g. telemetry, cell re-execution, etc.).

For this PR, I've essentially only changed code to better match our coding style and
reorganized the structure of the files to be easier to read. In followups we'll start
making more dramatic changes, but I wanted to keep this change as NFC as possible.

MODULAR_ORIG_COMMIT_REV_ID: 17a61dd011f354723897640d1580ea18e5faf7c3