[Mojo][LLDB] Use the async streams for the cli repl (#12149)
I was using `llvm::outs()` and `llvm::errs()` but I should have used `GetAsyncOutputStream` and `GetAsyncErrorStream` because it's possible to change the fd of the debuggers's `stderr` and `stdout` but leave `llvm::outs()` unchanged.

Because of this I had to convert this code to use the private API, which is fine.

MODULAR_ORIG_COMMIT_REV_ID: f7dbaf83d59c7287fb3988fecb176536869cbd7b