[Mojo][LLDB] Create a communication channel for non stderr/stdout messages between the plugin and higher level tools (#12150)
Stacked PRs:
 * __->__#12150

### [Mojo][LLDB] Create a communication channel for non stderr/stdout
messages between the plugin and higher level tools

This makes the type system plugin a broadcaster of messages that can be
received by higher level tools like the jupyter kernel. The type system
is being used for this because it's bound to the target and has easy
access from everywhere in the code, but there's no other particular
reason why we are piggybacking on it. There's also a
StructuredDataPlugin that we could use but it has some limitations (e.g.
if the macos structure data plugin is enabled, then our plugin would be
disabled because only one of those plug-ins can be active at a time per
process).

For additional context, I first tried reusing the process' stderr, but I
prefer this approach because it clearly separates the inferior's output
with these messages. In this case, the kernel is sending the fixit
messages to its stderr channel. These fixit messages are also being
printed to the debugger's stderr for completeness.

MODULAR_ORIG_COMMIT_REV_ID: 65bb339806e879877a555817b65eff75815a3dea