[KGEN][CheckLifetimes] Add dtor support for loops/break/continue. (#12232)
This enhances the CheckLifetimes pass to insert destructor calls
in the presence of hlcf.loop, and adds support for break/continue.

Break is quite easy because it just propagates the consume set from
the statement after the loop.  Continue is trickier: we have to
do iterative dataflow analysis on the body of the loop to determine
what the converged consume set needs to be, and (unlike uninit
checker) each pass side effects the IR by inserting destructors.

Address this by adding a 'dry mode' setting to the solver which
causes it to propagate and build the same sets of consumed values,
but does not actually insert the destructors.

Given these features, we can insert dtors into looping constructs
(even non-trivial ones) now!

MODULAR_ORIG_COMMIT_REV_ID: 160dbe632bfd8ab72d1f8509fbd01775b05f2cef