[mojo-lldb] Remove the disassembly logic from JitExecutionUnit (#12236)
This is only used for logging and is a large chunk of complex code
that we don't really need right now. When it becomes useful to
re-add, it will be trivial for Mojo (which already has all of this easily
accessible).

MODULAR_ORIG_COMMIT_REV_ID: 9978bb4c847838986c32f5c3f2df99c82ceaa3fb