[KGEN] Make the elaborator deal with RebindOp (#12252)
`kgen.rebind` is the main operation for bridging parameter domains
within the same parameter scope. It was handled in a sort of odd way by
KGEN. Its verifier rejects different concrete input and result types,
and we rely on post-elaboration folding to get rid of the concrete op.
This PR makes the elaborator explicitly handle the operation. There has
various benefits: relying on verifier errors is finnicky at the moment,
and we guarantee that this op is deleted during elaboration, but also we
want to allow IR constructs like:

```mlir
%0 = kgen.param.if <0> -> index {
  %none = kgen.param.constant: !lit.none = <#lit.none>
  %1 = kgen.rebind %none : !lit.none to index
  kgen.param.yield %1 : index
} else {
  %2 = kgen.call @make_index_value() : () -> index
  kgen.param.yield %2 : index
}
```

Where the `kgen.rebind` does not cause the IR as a whole to be rejected
because it resides inside a `kgen.param.if`. This will, among other
things, allow us to partially specialize IR.

MODULAR_ORIG_COMMIT_REV_ID: 462382e28ac1eef2f33ec842caa9700e662b0b54