[mojo-lldb] Use Event API for logging, NFC. (#12279)
This patch uses the event API for logging debug and error messages. This
enables us to be more selective about what we log, and when, to avoid
spam.

This patch sets a default size of 10 for the debug message cache - and
uses the same code to handle MojoTypeSystem events everywhere so this
policy is easily changed. Another option would be to log different
things to different files but log them immediately - all these are
policy decisions that we can decouple from the implementation once we
have clearer requirements.

MODULAR_ORIG_COMMIT_REV_ID: 8ab53454e3040cd841b2a5a1b6bdee3610f9912e