[KGEN] Add a LiftAndFoldApply pass (#12375)
This pass runs before elaboration and its goal is to lift 'apply'
operators out from parameter expressions into a new `kgen.param.apply`
"call" operation. This allows the elaborator to use the operation as a
"save point" if the callee happens to be multi-versioned, which is less
trivial to do during a parameter rewrite walk.

This pass will also fold lingering 'apply' operators with constant
inputs that escape the parser / occur due to inlining and other
processing, but that's in a follow-up.

MODULAR_ORIG_COMMIT_REV_ID: 73056ca707754cbf5035022b02392074a6390195