[mojo-lang][mojo-stdlib] Extend variadics to work with memory types/ownership. (#12431)
This enhances variadic lowering of memory-only types to make sure
the lifetime of the value in the variadic pack extends across the
function call that uses the pack.  It does this by introducing new
lit.ownership.use operations after the call that keep the arguments
live until then.

This also introduces a new `VariadicListMem` type, which is like
VariadicList but that works around the problem where !kgen.type
arguments cannot bind to memory only types or types with non-trivial
ownership (because it doesn't know how to copy and destroy things
in the absence of traits).  It is awkward to use, but provides a
projection of the raw pointer that can be used by the client
to get the value using `__get_address_as_lvalue`.

This then goes to show it off by extending `String.join` to be
properly variadic again, fixing a workaround I introduced last
night.

This fixes #12373

BUFFER_ORIG_REV_ID: 23f3731dedf7a2e19de9e7afc2b1616fd2fe1639