[Kernels] Simplify matmul elementwise epilogue contract and lift check to compile time (#12452)
Given the matmul structure the elementwise epilogue function should only
be called on the last K tile.

This change shifts the burden of checking when to call the epilogue from
the epilogue author to the matmul author, which allows the check to be
performed at compile time. Otherwise the same check needs to be
dynamically implemented in each epilogue, which is bad for perf, and the
author of the epilogue needs to know about matmul impl details.

LINALG_ORIG_REV_ID: d700e17f9fddc2820422552517f8f8881732abb7