[mojo-lang] Add slice syntax support (#12443)
Slice syntax `[a:b:c]` emits a call to the `slice` object constructor.
Because any of the slice indices may be `None`, the constructor in
question is made to be parametric on the index types. Right now it only
handles `NoneType` vs `Int`, but when we get traits, we can add handling
for invoking the `__index__` method on types.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: ef36171138e4260e76017f4e864a14f2bb482f89