[mojo-lldb] Add support for executing python expressions within the repl (#12399)
When the first line of the incoming expression matches `>python`, we treat
that expression as python and execute it using the python bindings within
the mojo process. The top-level python symbols are then imported back
into the mojo process, allowing follow-on expressions to reference the
python symbols seamlessly. Eventually, we'll also want to support the
reverse, i.e. exposing mojo to python seamlessly, but the bindings need
to be built out a bit more first.

The one part about this that is a bit limited is the detection of top-level
python symbols. For now we use a few simple regexs to detect common
situations, but realistically we should query pythons ast functionality to
determine the symbols to import. To do that we need lldb to be built
with python support, which makes this much easier, until then the simple
checks should cover the trivial cases we are interested in.

MODULAR_ORIG_COMMIT_REV_ID: 73c48ee4a3cf6929884e574d85b196ec6b0a9151