[KGEN] Fix a subtle bug in ExecutionEngine::create. (#12481)
River noticed that in #12189, the mutex offset is just above 0 - usually
indicating an offset from `nullptr`. That led to an `addToSearchOrder`
call with a JITDylib lookup (of the platform stdlib) that isn't created
in every code path.

The fix here is to essentially just always create the platform stdlib.
We always expect it to be there, and we register functions for the
debugger handlers and the orc_rt (when we have it), so we should just
create it unconditionally.

This should fix #12189.

MODULAR_ORIG_COMMIT_REV_ID: bcf7be82ea95e653a9c638a0030c46d22d92a127