[mojo-lldb] Flush the debug log on error state. (#12502)
If we log an error, we definitely want to flush the debug logs, so make
that behavior standard. Hold onto the function that explicitly flushes
for the special expression that we have.

This patch also increases the message cache size to 40 because we
generate about 8 messages per expression evaluated.

MODULAR_ORIG_COMMIT_REV_ID: ecac9439609da056d96428db81b24e79cf979ed7