[mojo-stdlib][mojo-lang] Object mega-PR (#12605)
The code in this stack is demo-critical, which means we cannot wait 7*20
minutes for them all to land. This mega-PR contains multiple PRs.

[mojo-stdlib] Add a `print` overload for `object`

And replace the `dump` method. This should've been done to begin with.

[mojo-stdlib] Cleanup the `object` implmentation

Added helpers to `_ObjectImpl` to reduce the amount of
`__get_address_as_lvalue` calls.

[mojo-stdlib] Document `Object.mojo`

Mikhail would be happy.

[mojo-stdlib] Add list comparators

Python lists are compared lexicographically.

[mojo-stdlib] Add "generic" object variant

This variant of object tries to model dynamic custom types with a
dictionary of attributes. This only supports setting and getting values
of object type, since we don't have a function variant for object.

MODULAR_ORIG_COMMIT_REV_ID: 037f0ce5eece3295ec898d52e35423c8b4382547