[mojo-jupyter] Horrible hack that ensures we bring in PythonObject's definition. (#12642)
This patch adds a function to the preamble of every mojo cell that
ensures we pull in the definition of PythonObject in every cell. This
should be removed post-haste as it is absolutely the incorrect solution
to this problem.

The underlying issue is roughly that we have persistent variables that
are of type kgen.declref<@"$PythonObject"::@PythonObject> which don't
then have the required definition. The right solution would be to
persist the actual AST nodes for these variables so we keep the structs
around.

MODULAR_ORIG_COMMIT_REV_ID: eb7fbd7ffdba5cced8ba2c801365cdc740bb771f