[mojo-lang] Fix memory safety hole with memory-only results. (#12672)
Abdul reported a problem where `x += y` and `x = x + y` were behaving
differently leading to a crash.  The problem is that the result of a
call (when the type is memory-only) is passed indirectly, and we
aggressively reuse buffers to avoid copies where possible.

This leads to a problem in cases like `x = foo(x)` or `x = x + y`
where the same LValue buffer is used for both the result slot and
the argument value.  It isn't correct to form concurrent mutable
and immutable references to the same buffer.  This would be rejected
by a borrow checker, but we don't have one yet.

Fix this by making the result slot optimization a "do it when known
safe" optimization based on simple alias analysis.  This enables us
to reuse the buffers in trivial cases, and the most important
situations like `let x = foo(y)` without running into weirdness.

The actual alias predicate is still trivial, I'll fill that in in
a subsequent patch.

MODULAR_ORIG_COMMIT_REV_ID: b6d55be68448af1c8cbb53b9be9671e3240007c6