[mojo-lang] Introduce consume operator `x^` (#12678)
This introduces the postfix consume operator, which produces an
rvalue given a lifetime tracked object (and, someday, a movable
lvalue).  This uses `lit.ownership.end.lifetime` to indicate the
end of a trackable lifetime and the start of a new one.

Right now we have a number of TODO's in the way of eliminating
memory-only copies, but this does help a lot with register-passable
values and is a step forward.

MODULAR_ORIG_COMMIT_REV_ID: da7bf9d327fce7bbf9bd9788d577158a3d91c673