[KGEN][CheckLifetimes] Model pop.store correctly. (#12687)
pop.store is used when storing an SRValue to memory, we were previously
handling the change to the stored slot correctly (running the dtor for
the overwritten value when needed) but we weren't modeling that store
consumes the value being stored.

We still have a problem with MRValue -> SRValue conversion as noted in
the fixme's.

This also moves some op comments into the description field so they
show up in the generated dialect docs.

MODULAR_ORIG_COMMIT_REV_ID: 8af6534b200b3c67457dde6e879d93b6073d0716