[Docs] Rename Mojo docs with website-style names and add webpage front matter (#12706)
The front-matter is YAML properties required by the static-website build
tool and is where we need to specify the page title (instead of as a
markdown heading). And all filenames for the website should be lowercase
with hyphens as spaces, instead of camelcase.

MODULAR_ORIG_COMMIT_REV_ID: e44f0d5c5e5b37f8ea1642576fe0625e8a0e1187