[mojo-lang] Implement ownership support for MRValue->SRValue conversion (#12727)
When referring to a `@register_passable` value in memory, a "load" needs
to
consume the input.  `POP::LoadOp` is a borrowing load, so we need to
use a new op to model this behavior.  This op is lowered to LowerLit
after ownership tracking is done.

MODULAR_ORIG_COMMIT_REV_ID: 1c0f47cb9a9206e0fc74c18150749bc960c1aafd