[KGEN] Make sure to erase nested scopes as well. (#12725)
When we process a param.if, we may have scopes under the region that we
don't process (and erase). In theory, simply erasing the regions for the
if-statement itself should work, given that we process them from the
inside-out, but what happens is that the region we *don't* process is
the one that has the nested parameters in it. That means we have to
erase any nested regions that are rooted on the if statement itself.

Ideally I'd like to ensure that nestedDecls and nestedScopes are in sync
(they currently are not always), but adding regions to nestedDecls
causes use-after-free issues.

MODULAR_ORIG_COMMIT_REV_ID: ea6e3d00c6002b5313b8b1c8a2c61f3f428f335e