[KGEN][CheckLifetimes] Properly model load borrows, fill in error message. (#12732)
A POP::LoadOp of a tracked value is a borrow of a value, so any use of
the load's value needs to extend the lifetime of the underlying value.

Similarly fill in a use-after-consume error message now that we have
consume.

MODULAR_ORIG_COMMIT_REV_ID: c9f0ed84765819ab730e8fa6a1ef3b50a24d2d34