[mojo-jupyter] Add `MOJO_JUPYTER_LOG_FILE` to control logging to a file. (#12739)
Apparently, docker containers have an issue where the stderr inside the container isn't always trace-able, and so the standard method is to log to a file specified in the environment. This hands control of the logging entirely to the environment, in hopes that our wonderful cloud folks can help us out of a jam wherein the stderr is not always piped to the container's log stream.

MODULAR_ORIG_COMMIT_REV_ID: 1e84ee41146106ae9223dfd1d933e99537ff1c3c