[API] Num-threads instead of config dict (#12815)
Use kwargs for Python API instead of a config dict.

This PR proposes we use
```python
session = InferenceSession(num_threads=2)
```
instead of

```python
session = InferenceSession({"thread_count": 2})
```

The former is more Pythonic, benefits from auto-completion in IDEs and
can be type-checked. The latter is prone to typos and relies on the
developer reading the documentation to discover potential keys for the
config dict like "thread_count".

Original discussion
[Internal link]
The main objection raised was the potential explosion of kwargs but this
is not an issue today.

MAX_PYTHON_ORIG_REV_ID: 5fcd5af64fec9eecb254c3dcd7cebdfa07100ef1