[mojo-lang] Further improvements to moveinit. (#12905)
This shifts responsibility for emitting OwnershipMarkDestroyedOp's
in `__del___` from LowerSemanticCF to the parser, by keeping track
of special function kinds in ASTDecl.  This allows us to simplify
a bunch of code, and start emitting them for the owned "existing"
argument in consuming move constructors written as
`fn __moveinit__(self&, owned existing: Self)`.

This in turn allows implementations of this method to take parts
out of the existing value, transfering ownership of those parts to
the new object without running the dtor on the existing one.  This
makes it an "actually useful" consuming move ctor.

MODULAR_ORIG_COMMIT_REV_ID: 1181f525609acc00fa9c6868832f796686f9e391