[KGEN] Concretize Capturing Regions into StageClosure Ops (#11976)
A materialization of a region with parameter values should result in a
new instance of a region that will later be lifted into a function. This
PR introduces the generation of those concrete regions with the
introduction of a StageClosure op.

MODULAR_ORIG_COMMIT_REV_ID: f2551bd089d6558f04af46bebe9f4557246bf4ab