[mojo-stdlib] Switch OwningOutputChainPtr to use automatic dtors. (#12997)
Automatic destructors are working, and is low hanging.

This is a reapply of #12964 with a fix to `AsyncTaskGroup.__del__`
(note: only two underscores).  This isn't an automatic del method,
so rename it to destroy() to remove ambiguity.  Since it isn't a
normal del method, it isn't running dtors on the elements.  We
manually destroy the out_chain in it.

MODULAR_ORIG_COMMIT_REV_ID: 55e3ec8392e8d16367e9c73866147d5c98726751