[mojo-lang][QoI] Fix location info in generated load. (#13001)
This fixes location information for a generated load in an arg
context to be the location of the value being loaded, not the
loc of the call.  Getting this wrong caused a poor error message.

This fixes Issue #12859.

MODULAR_ORIG_COMMIT_REV_ID: 00d56b9df69f859a59741a166a3b5a3daf2204d1