[mojo-kernel] Show a generic error message if the python code dies (#13004)
If python throws when evaluating an expression, the jupyter UI doesn't
show any indications of failures, which can be confusing. Instead, now
we are returning a response with status = "error" and printing a generic
message in the UI. The exception is still being printed in the stderr.

MODULAR_ORIG_COMMIT_REV_ID: 9e5bee0a0146dee3073fa62693320a53996cf296