[KGEN] Inline the Outlined Closures (#12882)
I did not use the Force Inline pass because the op that needs to be
replaced with the inlined closure is not a call but the
kgen.parameter.constant. It seemed inaccurate to make this op trigger
the callee dependency necessary for the inlining, so I just inlined in
the elaborator, which I know is slow.

Maybe this is okay knowing that it will go away once OutlineClosures
pass is eliminated and the elaborator is parallelized?

TODO:
- [ ] add more tests

MODULAR_ORIG_COMMIT_REV_ID: 922ac39858e9fa0a35f9461cf74c4d82d74cb07e