[mojo-lldb] Rework persistent variables to not reset on re-execution (#12928)
This PR works the support for persistent variables to be more
similar to python and other REPLs, in that it allows for naturally
supporting variable redefinition. This means that we don't have to
reset this and other execution state on re-execution of a previous
cell, and takes us a step further to removing the need to reset
anything.

MODULAR_ORIG_COMMIT_REV_ID: 264044f29a6b885bf0da952810d0c9332cbd53b7