[mojo-lldb][mojo-lang] Move the REPL source munging to the Parser library (#12929)
Right now the REPL has various interactions with Mojo source;
wrapping expressions, applying fixits, etc. Having this in the REPL was fine
for bootstrapping, but it's become a complex burden that we should cleanup
before the codebase grows too large that it's impossible. This PR fixes this
by moving all of the Mojo source touching pieces to the ParseLit library,
which now exposes a very thin API that the REPL utilizes when building
the expression to parse.

This has a vast number of benefits: it greatly simplifies the REPL codebase,
and creates a signficantly cleaner layering between the REPL and the mojo
parser. It also lets the parser solely manage all of the complexity of Mojo
source, meaning the REPL can focus on the pieces that it actually cares about.
This lets us completely remove any of the awkward state resetting, removes
the need for MojoExpressionSourceCode, and also strips quite a few other
pieces away from the REPL codebase.

While doing this, we naturally get various benefits for the REPL: we get
top-level code redefinition easily and correctly, meaning we can kill
the nasty temporary python hack we had to put in to force AST decls
to be alive across cells. It also more naturally allows for the REPL to
redefine state, making it significantly closer to a traditional python
like experience.

Aside from the above, this direction is also going to significantly
reduce the complexity of implementing the other LLDB functionality;
such as the type system methods (which want to interact with the
AST, but don't want to munge around all of the details that entails).
There are a lot of additional cleanups that this makes sooo much easier,
but I'm leaving those for followups to keep this as focused as possible
(given it's already large and touches a lot of things).

MODULAR_ORIG_COMMIT_REV_ID: 7100b08d2d5a424762f2699b0be41ce8b712e571