[mojo-lang] Introduce new `__take_pointee_as_owned_object(x)` builtin. (#13011)
The `__take_pointee_as_owned_object(x)` builtin takes a low-level
address value (of `!pop.pointer` type) and returns an `owned` value
for the memory that is pointed to.  This is an effective way to do a
"placement delete" on a pointer, or otherwise take ownership of the
pointed-to value.

This is extremely dangerous and should not be used by most code, but
is important for low-level hackery when we don't have full lifetimes
and traits available.

MODULAR_ORIG_COMMIT_REV_ID: cad7b7667b9be21d9ac9ff1cbbfc80c4a7442b8e