[mojo-lang] Implement indexing on variadics (#13018)
This PR implements IRGen for subscripting a variadic value like
`*a: Int` in parameter contexts (using a new parameter operator) and
dynamic contexts, for both registery and memory primary types.

This is not the ideal solution. The ideal solution would be for the type
of variadics to be emitted as VariadicList (and VariadicListMem), but
due to how pervasively VariadicType is used throughout the parser, this
is not possible at the moment (#13015).

MODULAR_ORIG_COMMIT_REV_ID: 97cb25e2337f8859973e815e81d0fbea2de8e448