[mojo-docs] Rewrite doc generation to produce json instead of markdown (#13022)
We currently generate markdown files for the docs, but this is really annoying
to maintain because styling involves munging around language implementation
details (which shouldn't be necessary), and also locks us into a specific format.
Generating JSON is significantly easier to manage, and creates a nice separation
of concerns between extracting information from the source code and actually
processing it to be rendered.

For now I added a simple script to convert the JSON to markdown using jinja2,
but we can easily tweak this to something different if we want to.

MODULAR_ORIG_COMMIT_REV_ID: 25ddb7c6071863d38c3b4bb978e2c2feb2511453