[mojo-jupyter] Some improvents in the logging functionality (#13040)
- Stop having multiple message types per event. A problem with that is
that the `stringifyType` method only picks one. It's also very easy to
ensure we only support one kind at a time.
- Add a MOJO_REPL_VERBOSE_LOG env var that enables eDumpIR and eDebugLog
messages, which are extremely noisy. This issue
[Internal link] was
created because currently we are emitting so many logs on the CI for it
to be useful to random people. However, we want it enabled on Jupyter
Hub.

MODULAR_ORIG_COMMIT_REV_ID: ceae693cda81e67640d7338a3bc2cdad353ce704