[mojo-jupyter] Trigger JSON logs based on an env var (#13113)
JSON logs are not very readable to common developers, but it's useful to
the JupyterHub environment. Therefore I'm adding an env var that will
control this feature. If MOJO_JUPYTER_JSON_LOGS is not passed, then the
logs are printed in textual form, which is better for normal
development.

MODULAR_ORIG_COMMIT_REV_ID: e81f1132298404d60624834efca512417371c60c