[KGEN] Tweaks to CheckLifetimes / lit.varlet.decl. (#13125)
This patch is an almost-but-not-NFC patch that improves one error
message in CheckLifetimes when dealing with structs that have no
members.  It also adds an "is synthesized" bit to lit.varlet.decl,
to distinguish between user written let/var declarations and ones
that are generated for other reasons by the frontend.
Most of the patch is updating testcases.

This includes a few other minor changes to the stdlib, switching
some var->let for example, but nothing significant.

This is the mechanical stuff broken out of a larger and more
interesting patch.

MODULAR_ORIG_COMMIT_REV_ID: 02553846e299d30bb710f8c5c057038df18f0acf