[mojo-lldb] Set the correct site-packages folder (#13101)
Read the comment in the CPython.mojo file for some additional context.
In short, CPython is searching for python libs in the system's
installation even a virtualenv is being used, which breaks the
expections of users on `install_python_deps`. Because of that, the
matplotlib integration of the jupyter notebook fails unless matplotlib
is installed system-wide. This problem can happen with other packages as
well.
See the comment mentioned above for more information about the fix.

MODULAR_ORIG_COMMIT_REV_ID: 07fd1f7809918657c75f21419647163ce45ab8f9