[mojo-lang] Diagnose problems with inconsistent indentation. (#13163)
Mojo is an indentation sensitive language and it can be very confusing
when one gets a mix of tabs and spaces - it can look like code is
indented
properly but the compiler doesn't think so.

This patch implements a check when lexing line-starting tokens to make
sure
that indentation across lines is consistently spelled. If not, we
produce
an error (as in the testcase).  Python has the equivalent check.

This fixes #12818.

MODULAR_ORIG_COMMIT_REV_ID: 2bc63565d6ad93cb8ae7d9fc55e463ec050eb7ca